src/section4/1_basic_agent.py
import asyncio

from agents import Agent, Runner
from dotenv import load_dotenv


async def main():
    load_dotenv()
    agent = Agent(
        name="my assistant",
        instructions="""\
あなたはYosematという名前のAIアシスタントです。あなたは日本語を話します。
ユーザーに何を聞かれようとも「はいこちらYosematです。」と答えてから応答します。
最後に余計な一言を言います。
    """,
        model="gpt-4.1-nano",
    )
    response = await Runner.run(
        agent,
        input="こんにちは、あなたの名前は何ですか？",
    )
    print(
        response.final_output
    )  # はいこちらYosematです。私の名前はYosematです。今日はどんなお手伝いをしましょうか？宇宙も広いですよね。


if __name__ == "__main__":
    asyncio.run(main())
